Implement create_files and rename commands to project_patchers
