+++ lib/mix/tasks/surface/surface.init/project_patchers/catalogue.ex
@@ -1,8 +1,10 @@
-defmodule Mix.Tasks.Surface.Init.Commands.Catalogue do
-  alias Mix.Tasks.Surface.Init.Patcher
-  alias Mix.Tasks.Surface.Init.Patchers
+defmodule Mix.Tasks.Surface.Init.ProjectPatchers.Catalogue do
+  @moduledoc false
 
-  @behaviour Mix.Tasks.Surface.Init.Command
+  alias Mix.Tasks.Surface.Init.ProjectPatcher
+  alias Mix.Tasks.Surface.Init.FilePatchers
+
+  @behaviour Mix.Tasks.Surface.Init.ProjectPatcher
 
   @impl true
   def file_patchers(%{catalogue: true} = assigns) do
@@ -35,7 +37,7 @@ defmodule Mix.Tasks.Surface.Init.Commands.Catalogue do
     web_folder = web_module |> inspect() |> Macro.underscore()
     dest = Path.join(["priv/catalogue/", web_folder])
 
-    Patcher.create_files(assigns, [
+    ProjectPatcher.create_files(assigns, [
       {"demo/example01.ex", dest},
       {"demo/example02.ex", dest},
       {"demo/playground.ex", dest}
@@ -47,7 +49,7 @@ defmodule Mix.Tasks.Surface.Init.Commands.Catalogue do
   def add_surface_catalogue_to_mix_deps do
     %{
       name: "Add `surface_catalogue` dependency",
-      patch: &Patchers.MixExs.add_dep(&1, ":surface_catalogue", ~S("~> 0.4.0")),
+      patch: &FilePatchers.MixExs.add_dep(&1, ":surface_catalogue", ~S("~> 0.4.0")),
       update_deps: [:surface_catalogue],
       instructions: """
       Add `surface_catalogue` to the list of dependencies in `mix.exs`.
@@ -69,8 +71,8 @@ defmodule Mix.Tasks.Surface.Init.Commands.Catalogue do
     %{
       name: "Configure `elixirc_paths` for the catalogue",
       patch: [
-        &Patchers.MixExs.add_elixirc_paths_entry(&1, ":dev", ~S|["lib"] ++ catalogues()|, "catalogues()"),
-        &Patchers.MixExs.append_def(&1, "catalogues", """
+        &FilePatchers.MixExs.add_elixirc_paths_entry(&1, ":dev", ~S|["lib"] ++ catalogues()|, "catalogues()"),
+        &FilePatchers.MixExs.append_def(&1, "catalogues", """
           [
             "priv/catalogue"
           ]\
@@ -101,7 +103,7 @@ defmodule Mix.Tasks.Surface.Init.Commands.Catalogue do
     %{
       name: "Configure esbuild for the catalogue",
       patch:
-        &Patchers.Phoenix.add_esbuild_entry_to_config(&1, :catalogue, """
+        &FilePatchers.Phoenix.add_esbuild_entry_to_config(&1, :catalogue, """
         [
             args: ~w(../deps/surface_catalogue/assets/js/app.js --bundle --target=es2016 --minify --outdir=../priv/static/assets/catalogue),
             cd: Path.expand("../assets", __DIR__),
@@ -130,8 +132,8 @@ defmodule Mix.Tasks.Surface.Init.Commands.Catalogue do
     %{
       name: "Configure catalogue route",
       patch: [
-        &Patchers.Phoenix.add_import_to_router(&1, Surface.Catalogue.Router, web_module),
-        &Patchers.Phoenix.append_route(&1, "/catalogue", web_module, """
+        &FilePatchers.Phoenix.add_import_to_router(&1, Surface.Catalogue.Router, web_module),
+        &FilePatchers.Phoenix.append_route(&1, "/catalogue", web_module, """
         if Mix.env() == :dev do
           scope "/" do
             pipe_through :browser
@@ -165,7 +167,7 @@ defmodule Mix.Tasks.Surface.Init.Commands.Catalogue do
     %{
       name: "Update patterns in :reload_patterns to reload catalogue files",
       patch:
-        &Patchers.Phoenix.add_live_reload_pattern_to_endpoint_config(
+        &FilePatchers.Phoenix.add_live_reload_pattern_to_endpoint_config(
           &1,
           ~S|~r"priv/catalogue/.*(ex)$"|,
           "catalogue",
@@ -195,7 +197,7 @@ defmodule Mix.Tasks.Surface.Init.Commands.Catalogue do
       name: "Add esbuild watcher for :catalogue",
       file: "config/dev.exs",
       patch:
-        &Patchers.Phoenix.add_watcher_to_endpoint_config(
+        &FilePatchers.Phoenix.add_watcher_to_endpoint_config(
           &1,
           :esbuild,
           "{Esbuild, :install_and_run, [:catalogue, ~w(--sourcemap=inline --watch)]}",
