+++ lib/mix/tasks/surface/surface.init/project_patchers/formatter.ex
@@ -1,7 +1,9 @@
-defmodule Mix.Tasks.Surface.Init.Commands.Formatter do
-  alias Mix.Tasks.Surface.Init.Patchers
+defmodule Mix.Tasks.Surface.Init.ProjectPatchers.Formatter do
+  @moduledoc false
 
-  @behaviour Mix.Tasks.Surface.Init.Command
+  alias Mix.Tasks.Surface.Init.FilePatchers
+
+  @behaviour Mix.Tasks.Surface.Init.ProjectPatcher
 
   @impl true
   def file_patchers(%{formatter: true}) do
@@ -35,7 +37,7 @@ defmodule Mix.Tasks.Surface.Init.Commands.Formatter do
     %{
       name: "Add `surface_formatter` dependency",
       update_deps: [:surface_formatter],
-      patch: &Patchers.MixExs.add_dep(&1, ":surface_formatter", ~S("~> 0.6.0")),
+      patch: &FilePatchers.MixExs.add_dep(&1, ":surface_formatter", ~S("~> 0.6.0")),
       instructions: """
       Add `surface_formatter` to the list of dependencies in `mix.exs`.
 
@@ -55,7 +57,7 @@ defmodule Mix.Tasks.Surface.Init.Commands.Formatter do
   def add_surface_inputs_to_formatter_config() do
     %{
       name: "Add file extensions to :surface_inputs",
-      patch: &Patchers.Formatter.add_config(&1, :surface_inputs, ~S(["{lib,test}/**/*.{ex,exs,sface}"])),
+      patch: &FilePatchers.Formatter.add_config(&1, :surface_inputs, ~S(["{lib,test}/**/*.{ex,exs,sface}"])),
       instructions: """
       In case you'll be using `mix format`, make sure you add the required file patterns
       to your `.formatter.exs` file.
@@ -75,7 +77,7 @@ defmodule Mix.Tasks.Surface.Init.Commands.Formatter do
   def add_sface_files_to_inputs_in_formatter_config() do
     %{
       name: "Add sface files to :inputs",
-      patch: &Patchers.Formatter.add_input(&1, ~S("{lib,test}/**/*.sface")),
+      patch: &FilePatchers.Formatter.add_input(&1, ~S("{lib,test}/**/*.sface")),
       instructions: """
       In case you'll be using `mix format`, make sure you add the required file patterns
       to your `.formatter.exs` file.
@@ -95,7 +97,7 @@ defmodule Mix.Tasks.Surface.Init.Commands.Formatter do
   def add_surface_to_import_deps_in_formatter_config() do
     %{
       name: "Add :surface to :import_deps",
-      patch: &Patchers.Formatter.add_import_dep(&1, ":surface"),
+      patch: &FilePatchers.Formatter.add_import_dep(&1, ":surface"),
       instructions: """
       In case you'll be using `mix format`, make sure you add `:surface` to the `import_deps`
       configuration in your `.formatter.exs` file.
@@ -115,7 +117,7 @@ defmodule Mix.Tasks.Surface.Init.Commands.Formatter do
   def add_formatter_plugin_to_formatter_config() do
     %{
       name: "Add Surface.Formatter.Plugin to :plugins",
-      patch: &Patchers.Formatter.add_plugin(&1, "Surface.Formatter.Plugin"),
+      patch: &FilePatchers.Formatter.add_plugin(&1, "Surface.Formatter.Plugin"),
       instructions: """
       In case you'll be using `mix format`, make sure you add `Surface.Formatter.Plugin`
       to the `plugins` in your `.formatter.exs` file.
